+++ plugins/websocket_server/WebsocketServer.cc
@@ -206,6 +206,24 @@ bool WebsocketServer::Load(const tinyxml2::XMLElement *_elem)
   }
   igndbg << "Using port[" << port << "]\n";
 
+  std::string sslCertFile = "";
+  std::string sslPrivateKeyFile = "";
+  elem = _elem->FirstChildElement("ssl");
+  if (elem)
+  {
+    // Get the ssl cert file, if present.
+    const tinyxml2::XMLElement *certElem =
+      elem->FirstChildElement("cert_file");
+    if (certElem)
+      sslCertFile = certElem->GetText();
+
+    // Get the ssl private key file, if present.
+    const tinyxml2::XMLElement *keyElem =
+      elem->FirstChildElement("private_key_file");
+    if (keyElem)
+      sslPrivateKeyFile = keyElem->GetText();
+  }
+
   // All of the protocols handled by this websocket server.
   this->protocols.push_back(
     {
@@ -230,7 +248,6 @@ bool WebsocketServer::Load(const tinyxml2::XMLElement *_elem)
   // The terminator
   this->protocols.push_back({NULL, NULL, 0, 0, 0, 0 });
 
-
   // We will handle logging
   lws_set_log_level( 0, lwsl_emit_syslog);
 
@@ -239,9 +256,40 @@ bool WebsocketServer::Load(const tinyxml2::XMLElement *_elem)
   info.port = port;
   info.iface = NULL;
   info.protocols = &this->protocols[0];
-  // We are not using SSL right now
-  info.ssl_cert_filepath        = NULL;
-  info.ssl_private_key_filepath = NULL;
+
+  if (!sslCertFile.empty() && !sslPrivateKeyFile.empty())
+  {
+    // Fail if the certificate file cannot be opened.
+    if (!ignition::common::exists(sslCertFile))
+    {
+      ignerr << "SSL certificate file[" << sslCertFile
+        << "] does not exist. Quitting.\n";
+      return false;
+    }
+
+    // Fail if the private key file cannot be opened.
+    if (!ignition::common::exists(sslPrivateKeyFile))
+    {
+      ignerr << "SSL private key file[" << sslPrivateKeyFile
+        << "] does not exist. Quitting.\n";
+      return false;
+    }
+
+    // Store SSL configuration.
+    info.options = LWS_SERVER_OPTION_DO_SSL_GLOBAL_INIT;
+
+    info.ssl_cert_filepath = sslCertFile.c_str();
+    info.ssl_private_key_filepath = sslPrivateKeyFile.c_str();
+  }
+  else if (sslCertFile.empty() || sslPrivateKeyFile.empty())
+  {
+    ignwarn << "Partial SSL configuration specified. Please specify: "
+    << "\t<ssl>\n"
+    << "\t  <cert_file>PATH_TO_CERT_FILE</cert_file>\n"
+    << "\t  <private_key_file>PATH_TO_KEY_FILE</private_key_file>\n"
+    << "\t</ssl>.\n"
+    << "Continuing without SSL.\n";
+  }
 
   // keep alive time of  60 seconds
   info.ka_time = 60;